+++ apps/backend/apps/admin/src/problem/resolvers/index.ts
@@ -1,0 +1,5 @@
+export * from './problem.resolver'
+export * from './testcase.resolver'
+export * from './problem-tag.resolver'
+export * from './file.resolver'
+// export * from './shared-group.resolver'
