+++ apps/backend/apps/admin/src/problem/services/index.ts
@@ -1,0 +1,4 @@
+export * from './file.service'
+export * from './problem.service'
+export * from './tag.service'
+export * from './testcase.service'
